Extracted constant functions into Functions for reuse
