+++ rxlifecycle/src/main/java/com/trello/rxlifecycle/TakeUntilGenerator.java
@@ -3,7 +3,6 @@ package com.trello.rxlifecycle;
 import android.support.annotation.CheckResult;
 import android.support.annotation.NonNull;
 import rx.Observable;
-import rx.exceptions.Exceptions;
 import rx.functions.Func1;
 import rx.functions.Func2;
 
@@ -33,28 +32,9 @@ final class TakeUntilGenerator {
                     return lifecycleEvent.equals(bindUntilEvent);
                 }
             })
-            .onErrorReturn(RESUME_FUNCTION)
-            .takeFirst(SHOULD_COMPLETE);
+            .onErrorReturn(Functions.RESUME_FUNCTION)
+            .takeFirst(Functions.SHOULD_COMPLETE);
     }
-
-    private static final Func1<Throwable, Boolean> RESUME_FUNCTION = new Func1<Throwable, Boolean>() {
-        @Override
-        public Boolean call(Throwable throwable) {
-            if (throwable instanceof OutsideLifecycleException) {
-                return true;
-            }
-
-            Exceptions.propagate(throwable);
-            return false;
-        }
-    };
-
-    private static final Func1<Boolean, Boolean> SHOULD_COMPLETE = new Func1<Boolean, Boolean>() {
-        @Override
-        public Boolean call(Boolean shouldComplete) {
-            return shouldComplete;
-        }
-    };
 
     private TakeUntilGenerator() {
         throw new AssertionError("No instances!");
